g2-common: update and hack releasetools for plain nougat
In case we're building on a plain AOSP Android tree, the 'build' repo
isn't patched for writing full boot images to the MMC sector.
Instead it extracts the image and uses imgdiff to create a patch, then
it applies the patch to the MMC sector.
For locked bootloaders that's forbidden, some of them like LG and
Samsung ones will check the boot.img contents' SHA and compare. If they
diverge the device will refuse to boot and will return integrity fail
because the new contents the patch applies have a different SHA.

We can workaround this by not using --block which is not safe, make
changes in the python script (build repo) to force full writes to the
sector on when using --block or either just copying the full image from
the device tree hookers.
Let's pick the most safe one, the last option and the one that ensures
that ramdisk matches the kernel.

Signed-off-by: Ícaro Hoff <icarohoff@gmail.com>